+++ src/geowombat/detect/metrics.py
@@ -22,7 +22,6 @@ from pathlib import Path
 import geopandas as gpd
 import numpy as np
 import pandas as pd
-from shapely.geometry import Polygon
 
 
 # ---------------------------------------------------------------------------